Implement just enough to mimick btrfs fi df
This commit lays out the initial structure of the btrfs module, and
implements a few first things.

* __init__.py contains everything that immediately has to be available
  for any python program that does import btrfs. Currently this is a
  FileSystem class, which is the entry point for retrieving all
  information about a btrfs filesystem.
* ctree.py contains some constants as seen in btrfs-progs ctree.h, and
  it will contain most internal filesystem object structures
* ioctl.py contains all communication using the kernel ioctl interface
  and related structs and low level plumbing tools. Functions in the
  ioctl module should return easily accessible python objects or named
  tuples or such to the caller.
* utils.py contains a bunch of helper utils that users of the btrfs
  module might not want to reimplement themselves.

The python code will not contain much documentation about low level data
structures and how they are related together. In order to understand
more of the code, I advise the reader to read the btrfs-progs and kernel
source code, as well as the btrfs wiki side by side with this code. I'll
try to keep naming of data structures etc. as much in sync with their
names in the kernel as possible, so that everything will be in a logical
place.

* examples/btrfs-fi-df.py contains a "clone" of the btrfs fi df
functionality, which demonstrates how easy it should be to get
information about your btrfs filesystem. Just skip through the 90%
command line argument parsing horror, to find all that is needed to
retrieve the data that it operates on:

    import btrfs
    spaces = btrfs.FileSystem('/').space_info()

spaces is a list of SpaceInfo named tuples, with fields named as one
would expect after looking up the ioctl_space_info struct in the kernel.

    # ./btrfs-fi-df.py /
    Data, single: total=14.94GiB, used=6.11GiB
    System, DUP: total=32.00MiB, used=16.00KiB
    Metadata, DUP: total=384.00MiB, used=150.34MiB
    GlobalReserve, single: total=64.00MiB, used=0.00B

Moo!