Search ioctl, search key, devices, chunks and stripes
A filesystem has devices and chunks, they're listed in the chunk tree.
In order to be able to list them, the search ioctl needs to be used.

This commit adds the search ioctl operation and handling of the search
key. The FileSystem object can now be queried for devices and chunks
with their stripes.

Searching for a block group in the extent tree can be done by using the
vaddr field from a chunk.

examples/show_usage.py lists all the physical locations of chunks for a
filesystem. By looking up the block group that the chunk belongs to, we
can show what percentage of a chunk is actually in use by data. Works
for single and RAID1 now, probably not for more fancy raid types yet.