Wasted space helper for RAID1: calculate unusable space
When using RAID1, it's possible to end up with unallocated space that is
"wasted", unusable.

When allocating a new block group, the chunk allocator will choose the
two devices with the most free space for the new mirrored chunks.

    >>> import btrfs
    >>> btrfs.utils.wasted_space_raid1([1,2,3,4])
    0
    >>> btrfs.utils.wasted_space_raid1([8,2,3,4])
    1
    >>> btrfs.utils.wasted_space_raid1([10,3])
    7
    >>> btrfs.utils.wasted_space_raid1([4,4,6,6])
    0

When calling this function with the list of device total sizes, the
result value is the amount of bytes that cannot be used at all, because
there is no data distribution possible to fill all the space. The most
simple example is a disk of size 1 and a disk of size 2. The wasted
space is 1.

When calling this function with the list of unallocated space on all
devices, the result value is the amount of bytes that cannot be used at
the moment, taking the current distribution of free space over the
devices in to account.

When combining these two, it's possible to find out which part of the
currently wasted space can probably be made usable again. By using btrfs
balance, the data can be redistributed a bit, so that the distribution
of free space over devices is optimal again.

In code:

    fs = btrfs.FileSystem('/')
    devices = list(fs.devices())
    device_total = [device.total_bytes for device in devices]
    device_unallocated = [device.total_bytes - device.bytes_used
                          for device in devices]
    wasted_total = btrfs.utils.wasted_space_raid1(device_unallocated)
    wasted_hard = btrfs.utils.wasted_space_raid1(device_total)
    wasted_soft = wasted_total - wasted_hard
    unallocated = sum(device_unallocated) - wasted_total

The "wasted_hard" is part of the unusable space that will never be
usable for RAID1 with the current device sizes. The "wasted_soft" means
space that can be made usable by running btrfs balance.

By calculating the unallocated space in this way, and using it for
monitoring, there will be less ENOSPC surprises.

Actually, this code is in the munin plugin that will be added in the
next commit.

By default, the helper uses 1G allocation steps, which can be overridden
using chunk_size.