Util function to get total fs usage
Since I'm already using this in several places, wht not put it in
utils...

    fs = btrfs.FileSystem('/')
    total, allocated, wasted_hard, wasted_soft, used = \
        btrfs.utils.fs_usage(fs)

* Total bytes (sum of the size of all attached devices)
* Allocated bytes (sum of the size all allocated chunks)
* "Wasted" bytes, see commit cf4089c for more information
* Used space, actual used space inside allocated space

The wasted space is a bit of a guestimation, because it shows what will
happen when filling up all available space with RAID0 or RAID1 data /
metadata, and assumes this is going to happen based on the occurance of
any RAID0 or RAID1 data chunk.

So, depending on the usefulness in your particular scenario, do
something with it, or ignore it, and just calculate unallocated space as
total minus used.