Show fat and skinny metadata extents
The Btrfs trees are stored as metadata inside extents. The extents used
for metadata are tracked inside the metadata. Crazy!

Extents could already been shown since a66853e "Show extents inside a
block group". This was extended with showing backreferences for data
extents in 0d928b4 "Show data extent backrefs".

This commit adds handling of the metadata tree block information and
backreferences.

There are two different ways this information can be stored:
* The original way, using a tree_block_info structure inside the extent
  item while setting flags to EXTENT_FLAG_TREE_BLOCK.
* The 'skinny' way, which is a more compact way of encoding all
  necessary information, which does not include the tree_block_info
  struct, and reuses the offset field in the object key for a different
  purpose. See kernel commit 3173a18 for more information about this.

The skinny metadata has its own key type: METADATA_ITEM_KEY.
In order to try keeping the code a bit understandable, I made the
MetaDataItem a separate class, instead of putting more if statements
inside the Extent class itself.

So, if the extent has flags DATA, it can have extent_data_refs and
shared_data_refs, which are inline, or separate objects following the
extent item itself, if there are too many of them to fit inline.

For fat metadata, the key type is also EXTENT_ITEM, but it has the flag
TREE_BLOCK set, because the extent contains btree data. In that case, it
contains the tree_block_info structure and inline refs:

    extent vaddr 85171486720 length 16384 refs 1 gen 1106551 flags TREE_BLOCK
        tree block key (18446744073709551606 EXTENT_CSUM 84375343104) level 0
        tree block backref root 7

btrfs-debug-tree shows this as:

    item 55 key (85171486720 EXTENT_ITEM 16384) itemoff 13427 itemsize 51
        extent refs 1 gen 1106551 flags TREE_BLOCK
        tree block key (EXTENT_CSUM EXTENT_CSUM 84375343104) level 0
        tree block backref root 7

For skinny metadata, the key type is METADATA_ITEM, and the tree level
and root backref is stored in the MetaDataItem object itself.

    metadata vaddr 122847232 refs 5 gen 91 flags TREE_BLOCK|FULL_BACKREF skinny level 0
        tree block backref root 579
        tree block backref root 573
        tree block backref root 567
        shared block backref parent 674381824
        shared block backref parent 130400256

btrfs-debug-tree shows this as:

    item 47 key (122847232 METADATA_ITEM 0) itemoff 11414 itemsize 69
        extent refs 5 gen 91 flags TREE_BLOCK|FULL_BACKREF
        tree block skinny level 0
        tree block backref root 579
        tree block backref root 573
        tree block backref root 567
        shared block backref parent 674381824
        shared block backref parent 130400256