Resolve extent vaddr to inodes using logical ino ioctl
The LOGICAL_INO ioctl, introduced in kernel commit 6055e736 can help us
converting an extent vaddr into a list of inode numbers of files that
use this extent.

    extent vaddr 87647895552 length 40960 refs 1 gen 1178414 flags DATA
        extent data backref root 5 objectid 4632126 offset 0 count 1
    extent vaddr 87685763072 length 20480 refs 1 gen 1106581 flags DATA
        shared data backref parent 87321182208 count 1

    >>> import btrfs
    >>> fs = btrfs.FileSystem('/')
    >>> btrfs.ioctl.logical_to_ino(fs.fd, 87647895552)
    ([Inode(inum=4632126, offset=0, root=5)], 0)
    >>> btrfs.ioctl.logical_to_ino(fs.fd, 87685763072)
    ([Inode(inum=3522197, offset=0, root=5)], 0)

For extents which have explicit backrefs, it doesn't look too magical,
but for the shared backrefs it's a bit more work to trace the metadata
backrefs and find files, which for example can reside in snapshotted
subvolumes.

In the fs tree of root 5, inum 3522197 can be found, which shows the
metadata pointing in the opposite direction. (not implemented in
python-btrfs yet, output from btrfs-debug-tree):

    item 36 key (3522197 INODE_ITEM 0) itemoff 13315 itemsize 160
            inode generation 936837 transid 936837 size 20451 nbytes 20480
            block group 0 mode 100644 links 1 uid 0 gid 0
            rdev 0 flags 0x10(none)
    item 37 key (3522197 INODE_REF 1459) itemoff 13295 itemsize 20
            inode ref index 2354 namelen 10 name: libwnck.mo
    item 38 key (3522197 EXTENT_DATA 0) itemoff 13242 itemsize 53
            extent data disk byte 87685763072 nr 20480
            extent data offset 0 nr 20480 ram 20480
            extent compression(none)

Instead of having a memory buffer inside the args (like the search ioctl does),
this ioctl expects a pointer to a separate buffer. Luckily the array module has
buffer_info(), which provides this information.

Currently, the default buffer size is 4096, which can hold max 170
results. The maximum size of the buffer that will be used to write the
results is 64kB.

The first 16 bytes of the returned information in the buffer are used
for a structure called btrfs_data_container, which contains a few bits
of information, like the amount of results returned, but also the amount
of extra bytes that would be needed if not all results fit into the
buffer.

An inode information structure (inum, offset, root) is 3x 8 bytes big,
which means that this ioctl can return a max of (65536 - 16) / 24 = 2730
results. So, if you have more than this amount of references to an
extent, you're out of luck listing them all in this way.

Let's see what happens when approaching the limits:

    -# btrfs sub create 0
    -# cp /some/file 0/0
    -# for i in $(seq 1 499); do cp --reflink 0/0 0/$i; done
    -# for i in $(seq 1 5); do btrfs sub snap 0 $i; done

This results in an extent which is used in 3000 files.

    -# ./show_block_groups.py /mnt/extents/
    block group vaddr 591396864 length 33554432 flags SYSTEM used 16384 used_pct 0
    block group vaddr 624951296 length 268435456 flags METADATA used 507904 used_pct 0
    block group vaddr 893386752 length 1073741824 flags DATA used 19869696 used_pct 2

    -# ./show_block_group_contents.py 893386752 /mnt/extents/
    block group vaddr 893386752 length 1073741824 flags DATA used 19869696 used_pct 2
    extent vaddr 913911808 length 19542016 refs 500 gen 3265 flags DATA
        extent data backref root 3461 objectid 778 offset 0 count 1
        extent data backref root 3461 objectid 759 offset 0 count 1
        extent data backref root 3461 objectid 777 offset 0 count 1
        extent data backref root 3461 objectid 780 offset 0 count 1
        extent data backref root 3461 objectid 783 offset 0 count 1
        extent data backref root 3461 objectid 763 offset 0 count 1
        extent data backref root 3461 objectid 789 offset 0 count 1
        extent data backref root 3461 objectid 774 offset 0 count 1
        [...]

Fun detail is that the extent has 500 backrefs instead of 3000. That
means that the snapshotting did not change this metadata. Just creating
one file in the subvolume and snapshotting it another 2999 times,
actually results in an extent with 3000 refs.

Anyway...

    -# python
    >>> import btrfs
    >>> fs = btrfs.FileSystem('/mnt/extents')
    >>> btrfs.ioctl.logical_to_ino(fs.fd, 913911808, 16)
    ([], 72000)

    >>> inodes, bytes_missed = btrfs.ioctl.logical_to_ino(fs.fd, 913911808, 65536)
    >>> len(inodes)
    2730
    >>> bytes_missed
    6480

A buffer of 16 bytes can hold 0 results, and it reports to need another
72000 bytes (3000 results)... But, the max that can be used is 64kB, so
we'll never see the last 270 results.

Next up: implementing data structures for the fs tree. I'm not sure yet
how to use this function, and Inode will probably also get another name.