Retrieve Device Stats
The ioctl to retrieve device stats was introduced in kernel commit
c11d2c23 and changed a bit in b27f7c0c1 shortly after that.

The call uses a buffer which is slightly larger than 1kB, because of
adding a field in the second commit, while not updating the padding.

The internal implementation is a bit overengineered, designed for
maximum flexibility and future proof forward and backward compatibility.
It never changed since it exists, so I'm just hardcoding the five
current stats that are returned for a device. As soon something gets
added in the future, it will be early enough to start making it a little
bit more flexible.

    -# ./fs_info.py /mnt/extents/
    max_id 2 num_devices 2 fsid f32900b2-d23d-46e8-95b5-b5bdf70de9b1
        nodesize 16384 sectorsize 4096 clone_alignment 4096
    devid 1 uuid 019dfc23-20c4-4c9f-970a-053c78624edd bytes_used 1375731712
        total_bytes 26843545600 path /dev/mapper/durka-extents
    devid 1 write_errs 0 read_errs 0 flush_errs 0 generation_errs 0 corruption_errs 0
    devid 2 uuid e2741672-3dd8-407e-aa11-a8686cb31f50 bytes_used 0
        total_bytes 10737418240 path /dev/mapper/durka-extents2
    devid 2 write_errs 0 read_errs 0 flush_errs 0 generation_errs 0 corruption_errs 0

Using the reset option, stats can be reset, just like btrfs dev stats -z
would do.