Retrieve subvolume info
This adds data structures for InodeItem and RootItem

A subvolume is actually a separate filesystem root. In the C code, this
is represented by a root_item struct. A root_item struct contains an
inode_item struct and quite some fields, listing the subvolume uuid, a
possible parent, uuids from subvolumes that it was send/received from,
etc...

The subvolumes() function on a FileSystem object returns a generator
that walks over the subvolume objects, retrieving them from the kernel.

The little helper top_level() returns the RootItem object for the top
level subvolume (ID 5), which is by default not included in the result
of subvolumes().

There's quite some attributes inside InodeItem and RootItem. I don't know what
all of them do yet, but we'll probably use them later somewhere.

About the cur/next logic in FileSystem.subvolumes: at any time, there
may be more than 1 item for a subvolume in the root tree. If the top
level btree block of a subvolume gets cowed, then first a new root item
with a higher generation number than the existing one gets written, then
the transaction gets committed, and afterwards the previous one gets
removed.

E.g. the root tree kan have items like (two times 260):

    (257 ROOT_ITEM 0)  -> btrfs sub create foo (actually gen 11)
    (258 ROOT_ITEM 11) -> btrfs sub snap -r foo foo-snap
    (260 ROOT_ITEM 11) -> btrfs sub snap foo bar
    (260 ROOT_ITEM 12) -> just got changed during a transaction commit

See "Btree Roots" and "Copy on Write Logging" in the btrfs design part
of the wiki for more explanation about this:
  https://btrfs.wiki.kernel.org/index.php/Btrfs_design#Btree_Roots
  https://btrfs.wiki.kernel.org/index.php/Btrfs_design#Copy_on_Write_Logging

Instead of writing a tedious example script with all the fields, we can
also use pprint to quickly dump the contents of an object:

    -# btrfs sub list -u /mnt/i320/fstree/
    ID 257 gen 13 top level 5 uuid a647140f-18b6-764b-b91a-49cb0d85b2c6 path foo
    ID 258 gen 11 top level 5 uuid 92c14cdc-0090-9c4f-b609-f4b52ad8b706 path foo-snap
    ID 260 gen 13 top level 5 uuid 71d58c25-9829-3d46-a720-f87a8e8e4411 path bar

    -# python
    >>> import btrfs
    >>> fs = btrfs.FileSystem('/mnt/i320/fstree')
    >>> subvolumes = list(fs.subvolumes())
    >>> import pprint

Here's the RootItem contents...

    >>> pprint.pprint(subvolumes[0].__dict__)
    {'byte_limit': 0,
     'bytenr': 4751360,
     'bytes_used': 16384,
     'ctime': <btrfs.ctree.TimeSpec object at 0x7fde2e4babd0>,
     'ctransid': 12,
     'dirid': 256,
     'drop_level': 0,
     'drop_progress': <btrfs.ctree.DiskKey object at 0x7fde2e4baad0>,
     'flags': 0,
     'generation': 13,
     'generation_v2': 13,
     'inode': <btrfs.ctree.InodeItem object at 0x7fde2e4ba990>,
     'key': <btrfs.ctree.Key object at 0x7fde2e4ba950>,
     'last_snapshot': 13,
     'level': 0,
     'otime': <btrfs.ctree.TimeSpec object at 0x7fde2e4bac10>,
     'otransid': 10,
     'parent_uuid': UUID('00000000-0000-0000-0000-000000000000'),
     'received_uuid': UUID('00000000-0000-0000-0000-000000000000'),
     'refs': 1,
     'rtime': <btrfs.ctree.TimeSpec object at 0x7fde2e4bac90>,
     'rtransid': 0,
     'stime': <btrfs.ctree.TimeSpec object at 0x7fde2e4bac50>,
     'stransid': 0,
     'uuid': UUID('a647140f-18b6-764b-b91a-49cb0d85b2c6')}

...the btree Key of the root item...

    >>> pprint.pprint(subvolumes[0].key.__dict__)
    {'_key': 1216083156315228480733184L,
     '_objectid': 257,
     '_offset': 0,
     '_type': 132}

...one of the many timestamps...

    >>> pprint.pprint(subvolumes[0].otime.__dict__)
    {'nsec': 964809675, 'sec': 1478722154}

...and here's the InodeItem:

    >>> pprint.pprint(subvolumes[0].inode.__dict__)
    {'atime': <btrfs.ctree.TimeSpec object at 0x7f1389f9ba10>,
     'block_group': 0,
     'ctime': <btrfs.ctree.TimeSpec object at 0x7f1389f9ba50>,
     'flags': 18446744071562067968L,
     'generation': 1,
     'gid': 0,
     'key': None,
     'mode': 16877,
     'mtime': <btrfs.ctree.TimeSpec object at 0x7f1389f9ba90>,
     'nbytes': 16384,
     'nlink': 1,
     'otime': <btrfs.ctree.TimeSpec object at 0x7f1389f9bad0>,
     'rdev': 0,
     'sequence': 0,
     'size': 3,
     'transid': 0,
     'uid': 0}

etc....

This could be made more pretty by implementing __repr__ for the ctree
objects, but that also requires that the generated code can be used to
feed back to python to instantiate the objects, which, currently, does
not make any sense, since we're only handling information that we got
from an external source, the kernel. Changing it or creating objects by
ourselves is a bit meaningless.