Allow to subtract from a ctree.Key
See examples/show_extent_tree_keys about when this can be convenient.

By the way: When adding 1 to the maximum key, it will wrap around to
zero again.  OTOH, when subtracting 1 from the minimum key, you'll end
up with the maximum one again.

>>> import btrfs
>>> key = btrfs.ctree.Key(0, 0, 0)
>>> str(key - 1)
'(-1 255 18446744073709551615)'
>>> key = btrfs.ctree.Key(-1, 255, 18446744073709551615)
>>> str(key + 1)
'(0 0 0)'

I don't really know if this should be considered a bug or a feature, but
for now it's a feature.