Don't load extent refs by default
While it's interesting to be able to see the information about
backreferences, they're not really useful in userspace.

When cowing a filesystem tree (for example, when taking a snapshot of a
subvolume with a large filesystem tree) only the top level block of the
filesystem tree is cowed.

As long as no changes are made on a specific lower level metadata tree
block, it will stay shared between the two trees.

This means that a lower level metadata tree block can contain
information like this...

    extent vaddr 2289631232 length 65536 refs 1 gen 8 flags DATA
        extent data backref root 5 objectid 260 offset 103350272 count 1

...but still have two backreferences...

    >>> import btrfs
    >>> fs = btrfs.FileSystem('.')
    >>> btrfs.ioctl.logical_to_ino(fs.fd, 2289631232)
    ([Inode(inum=260, offset=103350272, root=262),
     Inode(inum=260, offset=103350272, root=5)], 0)

...which can only be found when climbing the trees to see where the cow
happened, and following metadata tree block references in there.

Doing this from userspace is quite horrible, and some helpers like
the logical_ino ioctl exist that do this work in kernel for us.

Always creating all objects for the refs in here is a waste of cpu and
memory, so only do it when the user explicitely tells us to.