The Search IOCTL and transids...
First of all, rename transid_{min,max} -> {min,max}_transid
Yay for consistency. I have no idea why I put them in like this before.
And, I didn't use them yet, so yolo style API change.

The min_transid and max_transid are an in-kernel filter for objects that
are returned.

A transid (transaction ID) is very much related to the concept of a
generation in btrfs.

Every time the filesystem does a transaction commit, this number is
increased. transid and gen numbers that are left all over the place tell
us when an object was created or when it was last changed.

The search header that is returned for each result (see namedtuple
SearchHeader) contains a transid field that tells us in which
transaction the metadata record was last touched.