Keep transid value for BlockGroupItem
As Sargeth pointed out on IRC, the transid value that we get back from
the search ioctl for a block group can be quite interesting:

22:49 < Sargeth> Would be great if you could expose that transid as
well. I'd like to know if there are underutilised block groups that
haven't been touched for a while, because that means that their free
space is sufficiently fragmented as to be useless.
22:49 < Knorrie> yes, this is quite interesting
22:49 < Sargeth> Those would be excellent targets for a maintenance
balance

It seems that the transid of the BLOCK_GROUP_ITEM itself always is set
to the transid of the latest thing that happened anywhere in the
metadata of this block group, so it's good info to see how writes are
progressing, and how the allocator is choosing block groups for writes.

I'm not sure about other uses of transid yet, so let's start small.

Since the default string output of a blockgroup is used a lot for
reviewing block group usage for balance purposes, put the transid in it
directly.

    -# ./show_block_groups.py / |sort -nk 6 |tail -n 10
    block group vaddr 257719009280 transid 118228 length 33554432 flags SYSTEM|DUP used 49152 used_pct 0
    block group vaddr 77330382848 transid 118229 length 1073741824 flags DATA used 1073741824 used_pct 100
    block group vaddr 78404124672 transid 118229 length 1073741824 flags DATA used 1073741824 used_pct 100
    block group vaddr 79477866496 transid 118229 length 1073741824 flags DATA used 1025339392 used_pct 95
    block group vaddr 257182138368 transid 118235 length 536870912 flags METADATA|DUP used 52248576 used_pct 10
    block group vaddr 1094713344 transid 118236 length 1073741824 flags DATA used 807677952 used_pct 75
    block group vaddr 255571525632 transid 118236 length 536870912 flags METADATA|DUP used 342687744 used_pct 64
    block group vaddr 256108396544 transid 118236 length 536870912 flags METADATA|DUP used 375701504 used_pct 70
    block group vaddr 256645267456 transid 118236 length 536870912 flags METADATA|DUP used 201162752 used_pct 37
    block group vaddr 49413095424 transid 118236 length 1073741824 flags DATA used 1022074880 used_pct 95