crc32c for doing dir_item offset hash
The dir_item tree object is an item which holds a crc32c of the filename
in the offset field. It's meant as a way to be able to quickly find
directory items by name.

Also, add some python 2 and 3 if/else code. I don't know if it's 100%
sufficient. If not, I'd gladly refer to the fact that it might be
useful, but that there's no implied warranty of fitness for a particular
purpose.

Add dir_item_name_hash.py which can easily show the key offset number
for any filename.

    -$ python2 ./dir_item_name_hash.py παρίσταταί
    273446259
    -$ python3 ./dir_item_name_hash.py παρίσταταί
    273446259