ctree: clarify usage of data and buf
When working with the byte arrays:
- data is item data for a single tree object
- buf is any byte array from which item data can be unpacked, and/or
  which data can be packed back into at a certain position.
- pos is used together with a buf, to point at the location in the buf
  where we should do a read or write operation.

Examples:

In the search ioctl function, a 4096 byte buf is used to retrieve search
headers and item data (code from ioctl.py):

    buf = create_buf(4096)
    [...]
    fcntl.ioctl(fd, IOC_TREE_SEARCH, buf)
    result_nr_items = ioctl_search_key.unpack_from(buf, 0)[9]
    [...]
    header = SearchHeader(*ioctl_search_header.unpack_from(buf, pos))
    pos += ioctl_search_header.size
    data = buf[pos:pos+header.len]
    yield((header, data))

So, data is the exact slice with data for a single object, and pos is
used to walk around in the buf.

When using the search ioctl, we get a search header and data back for
each object:

    tree = CHUNK_TREE_OBJECTID
    min_key=Key(DEV_ITEMS_OBJECTID, DEV_ITEM_KEY, 1)
    max_key=Key(DEV_ITEMS_OBJECTID, DEV_ITEM_KEY, ULLONG_MAX)
    for header, data in btrfs.ioctl.search(fd, tree, min_key, max_key):
        yield DevItem(header, data)

But, the python objects use buf and pos again, because they can also
load themselves out of a larger buf. (Doing so will be added real soon.)

    class DevItem(object):
        def __init__(self, header, buf, pos=0):
            [...]

In case of loading a single item from data, the DevItem initializer just
consideres data to be a buf with starting position 0.

So, rename data to buf in initializers and add missing pos everywhere.