Better handling of IOCTL differences per arch
It would be a bit sad if this nice library would be limited to usage on
amd64 like systems.

The logic in this change is copied from what I could find in linux
kernel sources, augmented with a fair share of educated guessing.

In include/uapi/asm-generic/ioctl.h, the IOC macros get defined. But,
some architectures define IOC constants to their own value.

Reflect this behaviour by using the python platform.machine() and act
accordingly. platform.machine() returns the same as uname -m does

A better way of doing this, which I'd like to implement when packaging
this for debian, would be to create a compiled extension, which looks at
kernel source files when building it, on any architecture supported.

< jcristau> Knorrie: make a compiled extension instead
< jcristau> uname -m is not how you do this kind of stuff
< Knorrie> so it gets built correctly for every binary package?
< jcristau> yes
< Knorrie> interesting idea!

Until we have that, keep in this workaround. OTOH, this covers like 99%
of all users, and it's still nice to have as a fallback option, to be
able to just clone the source and do some things with it.

Big thanks to Adam Borowski (kilobyte) for spending half a day helping
me with this. Meow!