Inode lookup ioctl
This implements the INO_LOOKUP ioctl call, which seems to serve a number
of different purposes.

First of all, let's create a small example filesystem to play with:

    /mnt/i320 13-# mkfs.btrfs -f /dev/i320/fstree
    /mnt/i320 13-# mount /dev/i320/fstree fstree/
    /mnt/i320 13-# cd fstree/

    /mnt/i320/fstree 13-# btrfs sub create subvol
    /mnt/i320/fstree 13-# touch subvol/file1
    /mnt/i320/fstree 13-# touch subvol/file2
    /mnt/i320/fstree 13-# mkdir dir
    /mnt/i320/fstree 13-# touch dir/file3
    /mnt/i320/fstree 13-# touch dir/file4
    /mnt/i320/fstree 13-# ln dir/file3 dir/also_file3

    /mnt/i320/fstree 13-# btrfs sub list .
    ID 257 gen 11 top level 5 path subvol

    /mnt/i320/fstree 13-# tree
    .
    ├── dir
    │   ├── also_file3    <- hard link to file3
    │   ├── file3
    │   └── file4
    └── subvol            <- subvolume 257
        ├── file1
        └── file2

    2 directories, 5 files

1. Find the containing subvol id for any file:

By just providing a file descriptor to any file or directory, the
returned values are the subvolume tree id in which the file is placed,
and the path to the root of that subvolume, inside the subvolume, which
is always an empty string, (duh):

    >>> fd = os.open('/mnt/i320/fstree/', os.O_RDONLY)
    >>> btrfs.ioctl.ino_lookup(fd)
    (5, b'')

    >>> fd = os.open('/mnt/i320/fstree/subvol/file1', os.O_RDONLY)
    >>> btrfs.ioctl.ino_lookup(fd)
    (257, b'')

What actually happens here, is ino_lookup(fd, treeid=0, objectid=256),
after filling in the default arguments. This means: give me the tree id
and the place of inode 256 (the subvol root) inside the subvolume that
is containing whatever fd points to.

2. Find the containing subvolume and relative path for any file:

If we have an inode number of a file, we can open a file descriptor to
the file, call ino_lookup with the inode number and as result we get the
subvolume id and the relative path for that inode inside it:

    /mnt/i320/fstree 13-# stat subvol/file1 | pcregrep -o 'Inode: \d+'
    Inode: 259

    >>> fd = os.open('/mnt/i320/fstree/subvol/file1', os.O_RDONLY)
    >>> btrfs.ioctl.ino_lookup(fd, objectid=259)
    (257, b'file1/')

If an inode has multiple paths pointing to it (hard links), only the
first one found is returned:

    /mnt/i320/fstree 13-# stat dir/file3 | pcregrep -o 'Inode: \d+'
    Inode: 260
    /mnt/i320/fstree 13-# stat dir/also_file3 | pcregrep -o 'Inode: \d+'
    Inode: 260

    >>> fd = os.open('/mnt/i320/fstree/', os.O_RDONLY)
    >>> btrfs.ioctl.ino_lookup(fd, objectid=260)
    (5, b'dir/file3/')

    /mnt/i320/fstree 13-# rm dir/file3

    >>> btrfs.ioctl.ino_lookup(fd, objectid=260)
    (5, b'dir/also_file3/')

Also, there's always a trailing slash, which is a bit weird. The fact
that the functionality got introduced with the search ioctl and initial
btrfs subvol list functionality (to help building full paths of
subvolume locations in the output) maybe explains it.

3. Find the relative path for any inode number in any tree

When specifying a non-zero treeid, the ioctl will look up the inode in
that tree, instead of the tree that is holding the file which the open
filedescriptor points to:

    /mnt/i320/fstree 13-# stat dir | pcregrep -o 'Inode: \d+'
    Inode: 257

    >>> fd = os.open('/mnt/i320/fstree/', os.O_RDONLY)
    >>> btrfs.ioctl.ino_lookup(fd, treeid=5, objectid=257)
    (5, b'dir/')

    >>> btrfs.ioctl.ino_lookup(fd, treeid=257, objectid=260)
    (257, b'file2/')