Data structures for file bookkeeping
This adds the remaining data structures for file listings:
FileExtentItem, InodeExtref.

The FileExtentItem (using EXTENT_DATA_KEY as type) contains information
per file about where actual data is stored, or it stores inline data
inside itself.

    -$ cp /boot/vmlinuz-4.9.0-1-amd64 .

    -# ./show_file.py vmlinuz-4.9.0-1-amd64
    filename /mnt/btrfs/bla/foo/vmlinuz-4.9.0-1-amd64 tree 259 inum 304
    inode generation 64 transid 64 size 4185640 nbytes 4186112
        block_group 0 mode 100644 links 1 uid 1000 gid 1000 rdev 0
        flags 0x0(none)
    inode ref index 27731 name utf-8 vmlinuz-4.9.0-1-amd64
    extent data generation 64 ram_bytes 4186112 compression none
        type regular disk_bytenr 5172330496 disk_num_bytes 4186112
        offset 0 num_bytes 4186112

The extent data fields can be a bit confusing:
* (disk_bytenr EXTENT_ITEM disk_num_bytes) is the tree key of the extent
  item storing the actual data.
* offset is the offset inside that extent where the data we need starts.
* num_bytes is the amount of bytes to be used from that offset onwards.

Remember that these numbers always be multiples of disk block sizes,
because that's how it gets cowed. We don't just use 1 or 2 bytes from
another extent.

Now when we overwrite some of the file, cow happens, and we can see that
the first 16kiB are to be read from a new extent, while the previous one
still exists. For the remainder of the data, that original extent is
still being used, from offset 16384. Notice that the first (old) 16kiB
is still there, occupying space, but effectively unreachable via
regular file access.

    -$ dd if=/dev/urandom bs=16K conv=notrunc of=vmlinuz-4.9.0-1-amd64
        count=1

    -# ./show_file.py vmlinuz-4.9.0-1-amd64
    filename /mnt/btrfs/bla/foo/vmlinuz-4.9.0-1-amd64 tree 259 inum 304
        inode generation 64 transid 68 size 4185640 nbytes 4186112
        block_group 0 mode 100644 links 1 uid 1000 gid 1000 rdev 0
        flags 0x0(none)
    inode ref index 27731 name utf-8 vmlinuz-4.9.0-1-amd64
    extent data generation 68 ram_bytes 16384 compression none
        type regular disk_bytenr 5178613760 disk_num_bytes 16384
        offset 0 num_bytes 16384
    extent data generation 64 ram_bytes 4186112 compression none
        type regular disk_bytenr 5172330496 disk_num_bytes 4186112
        offset 16384 num_bytes 4169728

Inline and compressed files show up like this:

    -# ./show_file.py ./show_file.py
    filename ./show_file.py tree 260 inum 5757399
    inode generation 1952255 transid 1952256 size 1262 nbytes 1262
        block_group 0 mode 100755 links 1 uid 1000 gid 1000 rdev 0
        flags 0x0(none)
    inode ref index 307 name utf-8 show_file.py
    extent data generation 1952256 ram_bytes 1262 compression lzo
        type inline inline_encoded_nbytes 692

When hardlinking files, multiple names are stored inside one inode ref
object.

    -$ touch a
    -$ ln a b
    -$ ln a c

    -# ./show_file.py a
    filename a tree 259 inum 302
    inode generation 52 transid 58 size 0 nbytes 0 block_group 0
        mode 100644 links 3 uid 1000 gid 1000 rdev 0 flags 0x0(none)
    inode ref index 20162 name utf-8 a
    inode ref index 20163 name utf-8 b
    inode ref index 25429 name utf-8 c

If the amount of text stored in filenames makes the inode ref object
exceed the size of how big an object can be that can stored inside a
metadata page, the remainder items are stored in separate inode_extref
objects.

    -$ for i in $(seq 1 2000); do ln a a$i; done

    -# ./show_file.py a
    filename /mnt/btrfs/bla/foo/a tree 259 inum 302
    inode generation 52 transid 61 size 0 nbytes 0 block_group 0
        mode 100644 links 2003 uid 1000 gid 1000 rdev 0 flags 0x0(none)
    inode ref index 20162 name utf-8 a
    inode ref index 20163 name utf-8 b
    inode ref index 25429 name utf-8 c
    inode ref index 25730 name utf-8 a1
    inode ref index 25731 name utf-8 a2
    inode ref index 25732 name utf-8 a3
    [...]
    inode ref index 26881 name utf-8 a1152
    inode ref index 26882 name utf-8 a1153
    inode extref parent_objectid 297 index 27046 name utf-8 a1317
    inode extref parent_objectid 297 index 26900 name utf-8 a1171
    inode extref parent_objectid 297 index 27033 name utf-8 a1304
    [...]

All inode refs are stored inside one big metadata object, which in this
case has search key (302 INODE_REF 297). 297 is actually a reference to
the inode number of the directory this file is stored in.

The inode extrefs are stored in separate objects, which have a crc32c of
the filename in their key offset, e.g. (302 INODE_EXTREF 8005066) for
the index 27046 name a1317 file.

Files can also have xattrs attached:

    -$ touch file_with_xattr
    -$ setfattr -n user.mekker -v blaat file_with_xattr

    -# ./show_file.py /mnt/btrfs/bla/foo/file_with_xattr
    filename /mnt/btrfs/bla/foo/file_with_xattr tree 259 inum 303
    inode generation 63 transid 63 size 0 nbytes 0 block_group 0
        mode 100644 links 1 uid 1000 gid 1000 rdev 0 flags 0x0(none)
    inode ref index 27730 name utf-8 file_with_xattr
    xattr item hash 1703744141 name utf-8 user.mekker data utf-8 blaat