Also unpack bitmaps when listing free space
I already created a helper to list all fragments of free space, but that
code still totally ignored free space encoded as bitmaps.

The data inside a free space bitmap item is a sequence of bytes. For
each byte, the bits inside represent a sectorsize amount of free space
if the bit is set to 1. The bit with lowest significance is the first
sector, etc.

    >>> data = bytearray('\x6f\x00\0f')
    >>> for extent in btrfs.free_space_tree.unpack_bitmap(0, 4096, data):
    ...     print(extent)
    ...
    FreeSpaceExtent(vaddr=0, length=16384)
    FreeSpaceExtent(vaddr=20480, length=8192)
    FreeSpaceExtent(vaddr=65536, length=16384)

For example, the first byte is 6f, or 01101111. This means that the
first 4 (reading the bits from right to left) sectors are free space,
after that one is occupied, then two free, etc...