Revert adding transid field to BlockGroupItem
This reverts commit 9e9ef9958

Back then I incorrectly thought (see commit message on ca2d21e7b) that
the transid value that is put in the header of each search result does
have a meaning related to the item itself, like the transaction in which
the item being returned was created or changed.

This is not true.

This transaction id only denotes when the metadata page that currently
contains the item got written the last time as result of a COW
operation. This could happen because any item in the metadata page has
to change, and another example is btrfs balance causing rewriting of
metadata pages.