Introduce ItemData super class
Inherit every class that is used as actual item data from ItemData.

First use of this is to put the code that initializes the key attribute
which holds the tree key of the object in a single place.

Note: Some object types can either be wrapped into another object (like
an InodeItem inside a RootItem). If they're not directly indexed, it's
allowed to provide an explicit None as header, which will be ignored.