Use memoryview for partial buffers
In commit ff117369 "clarify usage of data and buf", I introduced using
a buf and pos in many places, so that in the "future" I could write back
data after changing it inside the objects. While doing so, I didn't
stick to the 'design for today' rule, because I wasn't actually putting
in that new functionality at the same time.

By using python memoryview, we can create zero-copy writable slices of
bytearrays to give to other objects and functions to use. This is better
than the buf + pos method, because it prevents an object from reading
and writing outside of its own range. When trying, it will just explode
and show there's a bug, instead of starting to behave in weird and hard
to debug ways.

The first bug fixed is inside MetaDataItem, where a "while pos <
len(buf)" would happily continue reading outside of its own object data
when being provided with a larger buffer... :|

Also, use a new bytearray for every search done using the search ioctl,
so that there's no chance that code is modifying the buffer in place
after it gets reused in a different place.

When using results from the search ioctl, we never need to keep a
reference to the buffer after loading data, or write anything back, but
reusing the buffer would still lead to surprises. In this code, it's
loading the subvolume root_item list that delays yielding a RootItem
until it knows it's the latest one. If a new search gets done in the
meantime, the data is overwritten before it's used to build up the
RootItem object.