Improve DirItem handling for multiple sub items
The dir item family (including DIR_ITEM, XATTR_ITEM and DIR_INDEX) all
share the same struct, the dir_item. However all of them might use some
fields and ignore others.

Also, the data for a single indexed DIR_ITEM and XATTR_ITEM can contain
multiple dir_item structs appended after each other, if multiple names
crc32 to the same "hash", which is put in the tree key offset field.

When adding DirItem in commit 8d4e17c0 "Data structures for the
directory index", I created the DirItem object to be able to hold data
for one dir_item struct.

Doing so resulted in a situation where untangling the multiple stored
objects is left as exercise for the reader. The following code, as seen
previously in the show_directory_contents example, is what needed to be
done manually. I don't like this, because it's too easy to just try
loading the data and not realize that the extra items are being missed.

    elif header.type == btrfs.ctree.DIR_ITEM_KEY:
        pos = 0
        while pos < header.len:
            item = btrfs.ctree.DirItem(header, data, pos)
            print(item)
            pos += len(item)

The new DirItemList handles the multiple sub items and acts as a list to
be able to access the sub elements:

    elif header.type == btrfs.ctree.DIR_ITEM_KEY:
        dir_item_list = btrfs.ctree.DirItemList(header, data)
        print(dir_item_list)
        for dir_item in dir_item_list:
            print("    {}".format(dir_item))

I still kept an XAttrItemList and XAttrItem class for convenience
reasons, it behaves identical to DirItem, except for the str() output.

However, the DirIndex object can only contain one dir_item struct.
To keep the code simple, I split that one off from DirItem. It keeps
working without changes.

This is a backwards incompatible change. The changes in the example
scripts show how to correctly iterate over the individual item data.

    -$ cd kebab
    -$ touch mouton
    -$ setfattr -n user.foobar -v 123 mouton
    -$ setfattr -n user.WvG1c1Td -v qwerty mouton
    -$ setfattr -n user.J3__T_Km3dVsW_ -v hello mouton
    -$ setfattr -n user.mekker -v blaat mouton
    -$ ln mouton schaap

    -# ./show_directory_contents.py ../kebab
    directory ../kebab tree 1713 inum 256
    inode generation 2263595 transid 2263625 [...]
    inode ref index 0 name utf-8 ..
    dir item list hash 3077446054 size 1
        dir item location (257 INODE_ITEM 0) type FILE name utf-8 schaap
    dir item list hash 3786996654 size 1
        dir item location (257 INODE_ITEM 0) type FILE name utf-8 mouton
    dir index 2 location (257 INODE_ITEM 0) type FILE name utf-8 mouton
    dir index 6 location (257 INODE_ITEM 0) type FILE name utf-8 schaap

    -# ./show_file.py ../kebab/mouton
    filename ../kebab/mouton tree 1713 inum 257
    inode generation 2263596 transid 2263625 [...]
    inode ref index 2 name utf-8 mouton
    inode ref index 6 name utf-8 schaap
    xattr item list hash 751495445 size 3
        xattr item name utf-8 user.foobar data utf-8 123
        xattr item name utf-8 user.WvG1c1Td data utf-8 qwerty
        xattr item name utf-8 user.J3__T_Km3dVsW_ data utf-8 hello
    xattr item list hash 1703744141 size 1
        xattr item name utf-8 user.mekker data utf-8 blaat

In case multiple filenames are used that compute to the same crc32c
value, they will show up as multiple items in the dir item list.